Farming diffs at this point
