+++ frontend/src/APIClients/CommonAPIClient.ts
@@ -50,30 +50,7 @@ const getUserStatus = async (email: string): Promise<string> => {
   }
 };
 
-const createResident = async ({
-  initial,
-  roomNum,
-  dateJoined,
-  building,
-}: Resident): Promise<boolean> => {
-  try {
-    const bearerToken = `Bearer ${getLocalStorageObjProperty(
-      AUTHENTICATED_USER_KEY,
-      "accessToken",
-    )}`;
-    await baseAPIClient.post(
-      "/residents",
-      { initial, roomNum, dateJoined, building },
-      { headers: { Authorization: bearerToken } },
-    );
-    return true;
-  } catch (error) {
-    return false;
-  }
-};
-
 export default {
   inviteUser,
   isUserInvited: getUserStatus,
-  createResident,
 };
